Move xray middleware to separate class
